Reorganized session package
